Refactor the configuration module
